Fix generating code for:
- toString, to show empty collections.
- mutate to keep presence of empty collections.
